% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-analyzer.R
\name{DataAnalyzer}
\alias{DataAnalyzer}
\title{It is used to analyze text data}
\description{
It provides information on text files, such as number of lines
and number of words. It allows generating sample file from an input text
file. It displays bar plots showing word frequencies.
}
\details{
It provides a method that returns text file information. The text
file information includes total number of lines, max, min and mean line
length and file size. It also provides a method that takes random samples of
lines in an input text file. It provides a method that reads an input text
file containing token frequencies. It displays the most occuring tokens.
}
\section{Super class}{
\code{\link[wordpredictor:TextFileProcessor]{wordpredictor::TextFileProcessor}} -> \code{DataAnalyzer}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DataAnalyzer$new()}}
\item \href{#method-plot_data}{\code{DataAnalyzer$plot_data()}}
\item \href{#method-get_file_info}{\code{DataAnalyzer$get_file_info()}}
\item \href{#method-generate_data}{\code{DataAnalyzer$generate_data()}}
\item \href{#method-get_ngrams}{\code{DataAnalyzer$get_ngrams()}}
\item \href{#method-clone}{\code{DataAnalyzer$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="wordpredictor" data-topic="TextFileProcessor" data-id="display_msg">}\href{../../wordpredictor/html/TextFileProcessor.html#method-display_msg}{\code{wordpredictor::TextFileProcessor$display_msg()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="wordpredictor" data-topic="TextFileProcessor" data-id="format_size">}\href{../../wordpredictor/html/TextFileProcessor.html#method-format_size}{\code{wordpredictor::TextFileProcessor$format_size()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="wordpredictor" data-topic="TextFileProcessor" data-id="post_process">}\href{../../wordpredictor/html/TextFileProcessor.html#method-post_process}{\code{wordpredictor::TextFileProcessor$post_process()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="wordpredictor" data-topic="TextFileProcessor" data-id="pre_process">}\href{../../wordpredictor/html/TextFileProcessor.html#method-pre_process}{\code{wordpredictor::TextFileProcessor$pre_process()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="wordpredictor" data-topic="TextFileProcessor" data-id="process">}\href{../../wordpredictor/html/TextFileProcessor.html#method-process}{\code{wordpredictor::TextFileProcessor$process()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="wordpredictor" data-topic="TextFileProcessor" data-id="process_file">}\href{../../wordpredictor/html/TextFileProcessor.html#method-process_file}{\code{wordpredictor::TextFileProcessor$process_file()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="wordpredictor" data-topic="TextFileProcessor" data-id="read_file">}\href{../../wordpredictor/html/TextFileProcessor.html#method-read_file}{\code{wordpredictor::TextFileProcessor$read_file()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="wordpredictor" data-topic="TextFileProcessor" data-id="read_lines">}\href{../../wordpredictor/html/TextFileProcessor.html#method-read_lines}{\code{wordpredictor::TextFileProcessor$read_lines()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="wordpredictor" data-topic="TextFileProcessor" data-id="read_obj">}\href{../../wordpredictor/html/TextFileProcessor.html#method-read_obj}{\code{wordpredictor::TextFileProcessor$read_obj()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="wordpredictor" data-topic="TextFileProcessor" data-id="save_obj">}\href{../../wordpredictor/html/TextFileProcessor.html#method-save_obj}{\code{wordpredictor::TextFileProcessor$save_obj()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="wordpredictor" data-topic="TextFileProcessor" data-id="write_file">}\href{../../wordpredictor/html/TextFileProcessor.html#method-write_file}{\code{wordpredictor::TextFileProcessor$write_file()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
It initializes the current object. It is used to set the file name
and verbose options.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataAnalyzer$new(file_name = NULL, verbose = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file_name}}{The path to the input file.}

\item{\code{verbose}}{If progress information should be displayed.}

\item{\code{line_count}}{The number of lines to read at a time.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_data"></a>}}
\if{latex}{\out{\hypertarget{method-plot_data}{}}}
\subsection{Method \code{plot_data()}}{
It reads ngram token frequencies from an input text file. The ngram
frequencies are then displayed in a bar plot. The type of plot is
specified by the type option. 'top_features' displays the top n most
occuring tokens along with their frequencies. 'coverage' displays
the number of words along with their frequencies.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataAnalyzer$plot_data(opts)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{opts}}{The options for analyzing the data.
type -> The type of plot to display. The options are:
'top_features', 'coverage'.
n -> For 'top_features', it is the number of top most occuring
tokens.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_file_info"></a>}}
\if{latex}{\out{\hypertarget{method-get_file_info}{}}}
\subsection{Method \code{get_file_info()}}{
Generates and returns information about the given text files.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataAnalyzer$get_file_info(file_list = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file_list}}{The list of text files to check.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data frame containing the overall file statistics.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-generate_data"></a>}}
\if{latex}{\out{\hypertarget{method-generate_data}{}}}
\subsection{Method \code{generate_data()}}{
It generates training, testing and validation data sets
from the given input file. It first reads the file given as a
parameter to the current object. It generates random indexes for the
data. It partitions the data into training, testing and validation
sets, according to the given parameters. The files are named
train.txt, test.txt and va.txt. The files are saved to the given
output folder.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataAnalyzer$generate_data(dir, percs)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dir}}{The name of the output folder.}

\item{\code{percs}}{The size of the training, testing and validation sets.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_ngrams"></a>}}
\if{latex}{\out{\hypertarget{method-get_ngrams}{}}}
\subsection{Method \code{get_ngrams()}}{
Returns the given number of ngrams and their
frequencies. If the prefix parameter is not given, then the ngrams
are randomly chosen. Otherise ngrams starting with the given regular
expression are returned.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataAnalyzer$get_ngrams(fn, c = NULL, pre = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fn}}{The ngram file name.}

\item{\code{c}}{The number of ngrams to return.}

\item{\code{pre}}{The ngram prefix, given as a regular expression.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataAnalyzer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
