% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sampler.R
\name{DataSampler}
\alias{DataSampler}
\title{It allows generating data samples from text files.}
\description{
It provides a method for generating training, testing and
validation data sets from a given input text file. It also provides a method
for generating a sample file of given size or number of lines from an input
text file.
}
\section{Super class}{
\code{\link[wordpredictor:TextFileProcessor]{wordpredictor::TextFileProcessor}} -> \code{DataSampler}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DataSampler$new()}}
\item \href{#method-generate_sample}{\code{DataSampler$generate_sample()}}
\item \href{#method-generate_data}{\code{DataSampler$generate_data()}}
\item \href{#method-clone}{\code{DataSampler$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
It initializes the current object. It is used to set the
verbose option.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSampler$new(ddir = "./data", mdir = "./models", ve = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ddir}}{The data directory.}

\item{\code{mdir}}{The model directory.}

\item{\code{ve}}{If progress information should be displayed.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-generate_sample"></a>}}
\if{latex}{\out{\hypertarget{method-generate_sample}{}}}
\subsection{Method \code{generate_sample()}}{
Generates a sample file of given size from the given
input file. The file is optionally cleaned and saved.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSampler$generate_sample(fn = NULL, ss, ic, ir, t, is)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fn}}{The input file name. It is the short file name relative to
the ddir. If not given, then the file name is auto generated from
the type parameter.}

\item{\code{ss}}{The number of lines or proportion of lines to sample.}

\item{\code{ic}}{If the sample file should be cleaned.}

\item{\code{ir}}{If the sample file contents should be randomized.}

\item{\code{t}}{The type of sample. It can be:
\itemize{
\item \strong{tr}. training sample.
\item \strong{te}. testing sample.
\item \strong{va}. validation sample.
}}

\item{\code{is}}{If the sampled data should be saved to a file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-generate_data"></a>}}
\if{latex}{\out{\hypertarget{method-generate_data}{}}}
\subsection{Method \code{generate_data()}}{
It generates training, testing and validation data sets
from the given input file. It first reads the file given as a
parameter to the current object. It partitions the data into
training, testing and validation sets, according to the given
parameters. The files are named train.txt, test.txt and va.txt. The
files are saved to the given output folder.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSampler$generate_data(fn, dir, percs)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fn}}{The input file name. It should be relative to the ddir.}

\item{\code{dir}}{The name of the output folder.}

\item{\code{percs}}{The size of the training, testing and validation sets.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataSampler$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
