% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-predictor.R
\name{ModelPredictor}
\alias{ModelPredictor}
\title{It is used to evaluate the accuracy and performance of the model.}
\description{
It provides methods that perform extrinsic and intrinsic model
evaluation. It also provides methods for determining the memory and time
requirements for generating the model. It also provides a method for
determining how much memory is used by the final model.
}
\details{
It provides a method that performs intrinsic model evaluation based
on Perplexity. It also provides a method that performs extrinsic model
evalation based on accuracy. It provides a method for determining how much
memory and time is needed to generate a model for different input data sizes.
It provides a method for determining how much memory is needed by the final
model.
}
\section{Super class}{
\code{\link[wordpredictor:TextFileProcessor]{wordpredictor::TextFileProcessor}} -> \code{ModelPredictor}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ModelPredictor$new()}}
\item \href{#method-get_model}{\code{ModelPredictor$get_model()}}
\item \href{#method-calc_perplexity}{\code{ModelPredictor$calc_perplexity()}}
\item \href{#method-predict_word}{\code{ModelPredictor$predict_word()}}
\item \href{#method-get_word_prob}{\code{ModelPredictor$get_word_prob()}}
\item \href{#method-clone}{\code{ModelPredictor$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
It initializes the current object. It is used to set the
model file name and verbose options.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelPredictor$new(mf, ve = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mf}}{The model file name.}

\item{\code{ve}}{If progress information should be displayed.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_model"></a>}}
\if{latex}{\out{\hypertarget{method-get_model}{}}}
\subsection{Method \code{get_model()}}{
Returns the Model class object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelPredictor$get_model()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The Model class object is returned.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-calc_perplexity"></a>}}
\if{latex}{\out{\hypertarget{method-calc_perplexity}{}}}
\subsection{Method \code{calc_perplexity()}}{
The Perplexity for the given sentence is calculated. For
each word, the probability of the word given the previous words is
calculated. The probabilities are multiplied and then inverted. The
nth root of the result is the perplexity, where n is the number of
words in the sentence. If the stem_words tokenization option was
specified, then the previous words are converted to their stem.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelPredictor$calc_perplexity(words)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{words}}{The list of words.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The perplexity of the given list of words.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-predict_word"></a>}}
\if{latex}{\out{\hypertarget{method-predict_word}{}}}
\subsection{Method \code{predict_word()}}{
Predicts the new word given a list of 1, 2 or 3 previous
words. It checks the given n words in the transition probabilities
data. If there is a match, the top 3 next words with highest
probabilities are returned. If there is no match, then the last n-1
previous words are checked. This process is continued until the last
word is checked. If there is no match, then empty result is returned.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelPredictor$predict_word(words, count = 3, dc = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{words}}{A character vector of previous words or a single vector
containing the previous word text.}

\item{\code{count}}{The number of results to return.}

\item{\code{dc}}{A DataCleaner object. If it is given, then the given words
are cleaned. If the stem_words option was set in the TokenGenerator
object configuration for the current model, then the words are
converted to their stems.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The top 3 predicted words along with their probabilities.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_word_prob"></a>}}
\if{latex}{\out{\hypertarget{method-get_word_prob}{}}}
\subsection{Method \code{get_word_prob()}}{
Calculates the probability of the given word given the
previous model-1 words, where model is the maximum ngram number. It
looks up the probability of a word given n previous words. The
previous n words are converted to numeric hash using digest2int
function. The hash is looked up in a data frame of transition
probabilities. The word is converted to a number by checking its
position in a list of unique words. If the hash and the word position
were found, then the probability of the previous word and hash is
returned. If it was not found, then the hash of the n-1 previous
words is taken and the processed is repeated. If the data was not
found in the data frame, then the word probability is returned. This
is known as backoff. If the word probability could not be found then
the default probability is returned. The default probability is
calculated as 1/(N+V), Where N = number of words in corpus and V is
the number of dictionary words.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelPredictor$get_word_prob(word, pw)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{word}}{The word whoose probability is to be calculated.}

\item{\code{pw}}{The previous words.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The probability of the word given the previous words.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelPredictor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
