% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token-generator.R
\name{TokenGenerator}
\alias{TokenGenerator}
\title{It generates ngrams of given size from an input text file}
\description{
It generates ngram tokens along with their frequencies.
}
\details{
It provides a method for generating ngrams of given size. It saves
each ngram along with its frequency to a text file.
}
\section{Super class}{
\code{\link[wordpredictor:TextFileProcessor]{wordpredictor::TextFileProcessor}} -> \code{TokenGenerator}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{tg_opts}}{The options for the token generator obj.
n -> The ngram size.
save_ngrams -> If the ngram data should be saved.
min_freq -> All ngrams with frequency less than min_freq are
ignored.
line_count -> The number of lines to process at a time.
stem_words -> If words should be converted to their stem.
dir -> The dir where the output file should be saved.
format -> The format for the output. There are two options.
'plain' -> The data is stored in plain text.
'obj' -> The data is stored as a R obj.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TokenGenerator$new()}}
\item \href{#method-generate_tokens}{\code{TokenGenerator$generate_tokens()}}
\item \href{#method-clone}{\code{TokenGenerator$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="wordpredictor" data-topic="TextFileProcessor" data-id="display_msg">}\href{../../wordpredictor/html/TextFileProcessor.html#method-display_msg}{\code{wordpredictor::TextFileProcessor$display_msg()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="wordpredictor" data-topic="TextFileProcessor" data-id="format_size">}\href{../../wordpredictor/html/TextFileProcessor.html#method-format_size}{\code{wordpredictor::TextFileProcessor$format_size()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="wordpredictor" data-topic="TextFileProcessor" data-id="post_process">}\href{../../wordpredictor/html/TextFileProcessor.html#method-post_process}{\code{wordpredictor::TextFileProcessor$post_process()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="wordpredictor" data-topic="TextFileProcessor" data-id="pre_process">}\href{../../wordpredictor/html/TextFileProcessor.html#method-pre_process}{\code{wordpredictor::TextFileProcessor$pre_process()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="wordpredictor" data-topic="TextFileProcessor" data-id="process">}\href{../../wordpredictor/html/TextFileProcessor.html#method-process}{\code{wordpredictor::TextFileProcessor$process()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="wordpredictor" data-topic="TextFileProcessor" data-id="process_file">}\href{../../wordpredictor/html/TextFileProcessor.html#method-process_file}{\code{wordpredictor::TextFileProcessor$process_file()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="wordpredictor" data-topic="TextFileProcessor" data-id="read_file">}\href{../../wordpredictor/html/TextFileProcessor.html#method-read_file}{\code{wordpredictor::TextFileProcessor$read_file()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="wordpredictor" data-topic="TextFileProcessor" data-id="read_lines">}\href{../../wordpredictor/html/TextFileProcessor.html#method-read_lines}{\code{wordpredictor::TextFileProcessor$read_lines()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="wordpredictor" data-topic="TextFileProcessor" data-id="read_obj">}\href{../../wordpredictor/html/TextFileProcessor.html#method-read_obj}{\code{wordpredictor::TextFileProcessor$read_obj()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="wordpredictor" data-topic="TextFileProcessor" data-id="save_obj">}\href{../../wordpredictor/html/TextFileProcessor.html#method-save_obj}{\code{wordpredictor::TextFileProcessor$save_obj()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="wordpredictor" data-topic="TextFileProcessor" data-id="write_file">}\href{../../wordpredictor/html/TextFileProcessor.html#method-write_file}{\code{wordpredictor::TextFileProcessor$write_file()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
It initializes the current obj. It is used to set the file name
and verbose options.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TokenGenerator$new(
  file_name = "./data/models/validate-clean.txt",
  opts = self$tg_opts,
  verbose = 0
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file_name}}{The path to the input file.}

\item{\code{opts}}{The options for generating the ngram tokens.
n -> The ngram size.
save_ngrams -> If the ngram data should be saved.
min_freq -> All ngrams with frequency less than min_freq are
ignored.
line_count -> The number of lines to process at a time.
stem_words -> If words should be converted to their stem.
dir -> The dir where the output file should be saved.
format -> The format for the output. There are two options.
'plain' -> The data is stored in plain text.
'obj' -> The data is stored as a R obj.}

\item{\code{verbose}}{Indicates if progress information should be displayed.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-generate_tokens"></a>}}
\if{latex}{\out{\hypertarget{method-generate_tokens}{}}}
\subsection{Method \code{generate_tokens()}}{
It generates ngram tokens and their frequencies from the given file
name. The tokens may be saved to a text file as plain text or a R
obj.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TokenGenerator$generate_tokens()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The ngram tokens along with their frequencies.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TokenGenerator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
