% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tp-generator.R
\name{TPGenerator}
\alias{TPGenerator}
\title{It is used to generate transition probabilities for n-grams.}
\description{
It provides a method for generating transition probabilities for the given
n-gram size. It also provides a method for generating the combined transition
probabilities data for n-gram sizes from 1 to the given size. The combined
transition probabilities data can be used to implement back-off.
}
\details{
It provides a method for generating n-gram transition probabilities. It reads
ngram frequencies from an input text file that is generated by the
TokenGenerator class. It parses each ngram into a prefix, a next word and the
next word frequency. Maximum Likelihood count is used to generated the next
word probabilities. The n-gram prefix is converted to a numeric hash. The
next word is replaced with the position of the next word in the list of all
words. The transition probabilities data is stored in a file. Another method
is provided that combines the transition probabilities for n-grams of size 1
to the given size. The combined transition probabilities can be saved as a
data frame. This file may be regarded as a n-gram model. The name of the file
is model-n.RDS, where n is the n-gram number. By combining the transition
probabilities of n-grams, back-off may be used to evaluate word probabilities
or predict the next word.
}
\section{Super class}{
\code{\link[wordpredictor:TextFileProcessor]{wordpredictor::TextFileProcessor}} -> \code{TPGenerator}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TPGenerator$new()}}
\item \href{#method-generate_tp}{\code{TPGenerator$generate_tp()}}
\item \href{#method-generate_tp_for_n}{\code{TPGenerator$generate_tp_for_n()}}
\item \href{#method-clone}{\code{TPGenerator$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
It initializes the current obj. It is used to set the
transition probabilities options and verbose option.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPGenerator$new(opts = list(), ve = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{opts}}{The options for generating the transition probabilities.
\itemize{
\item \strong{save_tp}. If the data should be saved.
\item \strong{n}. The ngram size.
\item \strong{dir}. The directory containing the input and output files.
\item \strong{format}. The format for the output. There are two options.
\itemize{
\item \strong{plain}. The data is stored in plain text.
\item \strong{obj}. The data is stored as a R obj.
}
}}

\item{\code{ve}}{If progress information should be displayed.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-generate_tp"></a>}}
\if{latex}{\out{\hypertarget{method-generate_tp}{}}}
\subsection{Method \code{generate_tp()}}{
It combines the transition probabilities for n-grams
less than or equal to the given n-gram size. It first generates
the transition probabilities for each ngram size from 1 to the given
size. The transition probabilities are then combined into a single
data frame and saved to the output folder that is given as parameter
to the current object. By combining the transition probabilities for
all n-gram sizes less than n, back-off can be used to calculate next
word probabilities or predict the next word.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPGenerator$generate_tp()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-generate_tp_for_n"></a>}}
\if{latex}{\out{\hypertarget{method-generate_tp_for_n}{}}}
\subsection{Method \code{generate_tp_for_n()}}{
It reads ngram token frequencies from an input text
file. It generates a data frame containing the prefix, next word
and next word frequency. The data frame may be saved to a file as
plain text or as a R obj. For n = 1, the list of words is saved.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPGenerator$generate_tp_for_n(n)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{The n-gram size for which the tp data is generated.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TPGenerator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
