% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordpredictor-package.R
\docType{package}
\name{wordpredictor-package}
\alias{wordpredictor}
\alias{wordpredictor-package}
\title{wordpredictor: Develop Text Prediction Models Based on N-Grams}
\description{
The wordpredictor package allows developing n-gram models for text
    prediction. It provides methods for data cleaning, data sampling,
    tokenization, model generation, model evaluation and word prediction. For
    information on how n-gram models work we referred to: Jurafsky, Daniel &
    Martin, James. (2008). "Speech and Language Processing: An Introduction to
    Natural Language Processing, Computational Linguistics, and Speech
    Recognition" <https://web.stanford.edu/~jurafsky/slp3/ed3book_dec302020.pdf>.
    For optimizing R code and using R6 classes we reffered to Hadley Wickham,
    "Advanced R" <http://adv-r.had.co.nz/R5.html>. For writing R extensions we
    refer to Hadley Wickham and Jenny Bryan, "R Packages",
    <https://r-pkgs.org/index.html>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pakjiddat/word-predictor}
  \item \url{https://pakjiddat.github.io/word-predictor/}
  \item Report bugs at \url{https://github.com/pakjiddat/word-predictor/issues}
}

}
\author{
\strong{Maintainer}: Nadir Latif \email{pakjiddat@gmail.com} (\href{https://orcid.org/0000-0002-7543-7405}{ORCID})

}
\keyword{internal}
