% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-analyzer.R
\name{DataAnalyzer}
\alias{DataAnalyzer}
\title{It is used to analyze text data}
\description{
It provides information on text files, such as number of lines
and number of words. It allows generating sample file from an input text
file. It displays bar plots showing word frequencies.
}
\details{
It provides a method that returns text file information. The text
file information includes total number of lines, max, min and mean line
length and file size. It also provides a method that takes random samples of
lines in an input text file. It provides a method that reads an input text
file containing token frequencies. It displays the most occuring tokens.
}
\section{Super class}{
\code{\link[wordpredictor:TextFileProcessor]{wordpredictor::TextFileProcessor}} -> \code{DataAnalyzer}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DataAnalyzer$new()}}
\item \href{#method-plot_n_gram_stats}{\code{DataAnalyzer$plot_n_gram_stats()}}
\item \href{#method-get_file_info}{\code{DataAnalyzer$get_file_info()}}
\item \href{#method-get_ngrams}{\code{DataAnalyzer$get_ngrams()}}
\item \href{#method-clone}{\code{DataAnalyzer$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
It initializes the current object. It is used to set the file name
and verbose options.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataAnalyzer$new(fn = NULL, ve = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fn}}{The path to the input file.}

\item{\code{ve}}{If progress information should be displayed.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_n_gram_stats"></a>}}
\if{latex}{\out{\hypertarget{method-plot_n_gram_stats}{}}}
\subsection{Method \code{plot_n_gram_stats()}}{
It reads ngram token frequencies from an input text
file. The ngram frequencies are then displayed in a bar plot. The
type of plot is specified by the type option. 'top_features' displays
the top n most occuring tokens along with their frequencies.
'coverage' displays the number of words along with their frequencies.
The plot stats are returned by the function as a data frame.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataAnalyzer$plot_n_gram_stats(opts)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{opts}}{The options for analyzing the data.
type -> The type of plot to display. The options are:
'top_features', 'coverage'.
n -> For 'top_features', it is the number of top most occuring
tokens. For 'coverage' it is the first n frequencies.
save_to -> The graphics devices to save the plot to.
NULL implies plot is printed.
dir -> The output directory where the plot will be saved.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data frame containing the stats.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_file_info"></a>}}
\if{latex}{\out{\hypertarget{method-get_file_info}{}}}
\subsection{Method \code{get_file_info()}}{
Generates and returns information about text files.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataAnalyzer$get_file_info(res)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{res}}{The name of a directory or a file name.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data frame containing the file statistics.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_ngrams"></a>}}
\if{latex}{\out{\hypertarget{method-get_ngrams}{}}}
\subsection{Method \code{get_ngrams()}}{
Returns the given number of ngrams and their
frequencies. If the prefix parameter is not given, then the ngrams
are randomly chosen. Otherise ngrams starting with the given regular
expression are returned.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataAnalyzer$get_ngrams(fn, c = NULL, pre = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fn}}{The ngram file name.}

\item{\code{c}}{The number of ngrams to return.}

\item{\code{pre}}{The ngram prefix, given as a regular expression.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataAnalyzer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
