% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-cleaner.R
\name{DataCleaner}
\alias{DataCleaner}
\title{Provides data cleaning functionality}
\description{
It provides a memory efficient method for removing unneeded
characters from text files. It is suitable for cleaning large text files.
}
\details{
It provides a method for cleaning text files. It allows removing bad
words, stop words, non dictionary words, extra space, punctuation and
non-alphabet characters. Allows conversion to lower case. It supports large
text files.
}
\section{Super class}{
\code{\link[wordpredictor:TextFileProcessor]{wordpredictor::TextFileProcessor}} -> \code{DataCleaner}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DataCleaner$new()}}
\item \href{#method-clean_file}{\code{DataCleaner$clean_file()}}
\item \href{#method-clean_lines}{\code{DataCleaner$clean_lines()}}
\item \href{#method-clone}{\code{DataCleaner$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
It initializes the current object. It is used to set the file name
and verbose options.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataCleaner$new(fn = NULL, opts = list(), ve = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fn}}{The path to the file to clean.}

\item{\code{opts}}{The options for data cleaning.
\itemize{
\item \strong{min_words}. The minimum number of words per sentence.
\item \strong{line_count}. The number of lines to read and clean at a time.
\item \strong{save_data}. If the combined processed lines should be saved.
\item \strong{output_file}. Name of the output file used to store the data.
\item \strong{sw_file}. The stop words file path.
\item \strong{dict_file}. The dictionary file path.
\item \strong{bad_file}. The bad words file path.
\item \strong{to_lower}. If the words should be converted to lower case.
\item \strong{remove_stop}. If stop words should be removed.
\item \strong{remove_punct}. If punctuation symbols should be removed.
\item \strong{remove_non_dict}. If non dictionary words should be removed.
\item \strong{remove_non_alpha}. -> If non alphabet symbols should be removed.
\item \strong{remove_extra_space}. -> If leading, trailing and double spaces
should be removed.
\item \strong{remove_bad}. If bad words should be removed
}}

\item{\code{ve}}{Indicates if progress information should be displayed.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clean_file"></a>}}
\if{latex}{\out{\hypertarget{method-clean_file}{}}}
\subsection{Method \code{clean_file()}}{
It provides an efficient method for cleaning text files.
It removes unneeded characters from the given text file with several
options. It allows removing punctuations, bad words, stop words,
non-alphabetical symbols and non-dictionary words. It reads a certain
number of lines from the given text file. It removes unneeded
characters from the lines and then saves the lines to an output text
file. File cleaning progress is displayed if the verbose option was
set in the class constructor. It is suitable for cleaning large text
files.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataCleaner$clean_file()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clean_lines"></a>}}
\if{latex}{\out{\hypertarget{method-clean_lines}{}}}
\subsection{Method \code{clean_lines()}}{
It cleans the given lines of text using the options
passed to the current object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataCleaner$clean_lines(lines)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lines}}{The input sentences.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The cleaned lines of text.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataCleaner$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
