% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text-file-processor.R
\name{TextFileProcessor}
\alias{TextFileProcessor}
\title{Base class for text file processing}
\description{
Provides basic structure for processing text files.
}
\details{
It provides pre-processing, processing and post-processing functions, which
need to be overridden by derived classes. The pre-processing function is
called before reading a file. The process function is called for processing a
line. The post processing function is called on the processed data.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TextFileProcessor$new()}}
\item \href{#method-clone}{\code{TextFileProcessor$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
It initializes the current object. It is used to set the file name
and verbose options.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TextFileProcessor$new(fn = NULL, lc = 100, ve = 2)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fn}}{The path to the file to clean.}

\item{\code{lc}}{The number of lines to read and clean at a time.}

\item{\code{ve}}{Indicates if progress information should be displayed.
Reads the given file one line at a time. It runs the given
pre-processing function before reading the file. It runs the given}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TextFileProcessor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
