% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token-generator.R
\name{TokenGenerator}
\alias{TokenGenerator}
\title{It generates ngrams of given size from an input text file}
\description{
It generates ngram tokens along with their frequencies. The data
may be saved to a file in plain text format or as a R object.
}
\section{Super class}{
\code{\link[wordpredictor:TextFileProcessor]{wordpredictor::TextFileProcessor}} -> \code{TokenGenerator}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TokenGenerator$new()}}
\item \href{#method-generate_tokens}{\code{TokenGenerator$generate_tokens()}}
\item \href{#method-clone}{\code{TokenGenerator$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
It initializes the current obj. It is used to set the file name,
tokenization options and verbose option.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TokenGenerator$new(fn = NULL, opts = list(), ve = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fn}}{The path to the input file.}

\item{\code{opts}}{The options for generating the ngram tokens.
n -> The ngram size.
save_ngrams -> If the ngram data should be saved.
min_freq -> All ngrams with frequency less than min_freq are
ignored.
line_count -> The number of lines to process at a time.
dir -> The dir where the output file should be saved.
format -> The format for the output. There are two options.
'plain' -> The data is stored in plain text.
'obj' -> The data is stored as a R obj.}

\item{\code{ve}}{Indicates if progress information should be displayed.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-generate_tokens"></a>}}
\if{latex}{\out{\hypertarget{method-generate_tokens}{}}}
\subsection{Method \code{generate_tokens()}}{
It generates ngram tokens and their frequencies from the
given file name. The tokens may be saved to a text file as plain text
or a R object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TokenGenerator$generate_tokens()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The data frame containing ngram tokens along with their
frequencies.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TokenGenerator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
