% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordpredictor-package.R
\docType{package}
\name{wordpredictor-package}
\alias{wordpredictor}
\alias{wordpredictor-package}
\title{wordpredictor: Develop Text Prediction Models Based on N-Grams}
\description{
A framework for developing n-gram models for text prediction. It provides data cleaning, data sampling, extracting tokens from text, model generation, model evaluation and word prediction. For information on how n-gram models work we referred to: "Speech and Language Processing" \url{https://web.archive.org/web/20240919222934/https\%3A\%2F\%2Fweb.stanford.edu\%2F~jurafsky\%2Fslp3\%2F3.pdf}. For optimizing R code and using R6 classes we referred to "Advanced R" \url{https://adv-r.hadley.nz/r6.html}. For writing R extensions we referred to "R Packages", \url{https://r-pkgs.org/index.html}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pakjiddat/word-predictor}
  \item \url{https://pakjiddat.github.io/word-predictor/}
  \item Report bugs at \url{https://github.com/pakjiddat/word-predictor/issues}
}

}
\author{
\strong{Maintainer}: Nadir Latif \email{pakjiddat@gmail.com} (\href{https://orcid.org/0000-0002-7543-7405}{ORCID})

}
\keyword{internal}
